// mcp/src/toggl/generated.ts
export interface components {
    schemas: {
// ...
        "github_com_toggl_toggl_api_internal_models.Project": {
            /** @description Whether the project is active or archived */
            active?: boolean;
            /** @description Actual hours */
            actual_hours?: number | null;
            /** @description Actual seconds */
            actual_seconds?: number | null;
            /** @description Last updated date */
            at?: string;
            /** @description Whether estimates are based on task hours, premium feature */
            auto_estimates?: boolean | null;
            /** @description Whether the project is billable, premium feature */
            billable?: boolean | null;
            can_track_time?: boolean;
            /** @description Client ID legacy field */
            cid?: number;
            /** @description Client ID */
            client_id?: number | null;
            client_name?: string;
            /** @description Color */
            color?: string;
            /** @description Creation date */
            created_at?: string;
            /** @description Currency, premium feature */
            currency?: string | null;
            /** @description Current project period, premium feature */
            current_period?: components["schemas"]["models.RecurringPeriod"] | null;
            /** @description End date */
            end_date?: string;
            /** @description Estimated hours */
            estimated_hours?: number | null;
            /** @description Estimated seconds */
            estimated_seconds?: number | null;
            /** @description ExternalReference can be used to store an external reference to the Track Project Entity. */
            external_reference?: string;
            /** @description Fixed fee, premium feature */
            fixed_fee?: number;
            /** @description Project ID */
            id?: number;
            /** @description The external ID of the linked entity in the external system (e.g. JIRA/SalesForce) */
            integration_ext_id?: string;
            /** @description The external type of the linked entity in the external system (e.g. JIRA/SalesForce) */
            integration_ext_type?: string;
            /** @description The provider (e.g. JIRA/SalesForce) that has an entity linked to this Toggl Track entity */
            integration_provider?: components["schemas"]["models.IntegrationProvider"];
            /** @description Whether the project is private */
            is_private?: boolean;
            /** @description Name */
            name?: string;
            permissions?: string[];
            pinned?: boolean;
            /** @description Hourly rate */
            rate?: number;
            /** @description Last date for rate change */
            rate_last_updated?: string | null;
            /** @description Whether the project is recurring, premium feature */
            recurring?: boolean;
            /** @description Project recurring parameters, premium feature */
            recurring_parameters?: components["schemas"]["models.RecurringProjectParameters"][];
            /** @description Start date */
            start_date?: string;
            /** @description Status of the project (upcoming, active, ended, archived, deleted) */
            status?: components["schemas"]["models.ProjectStatus"];
            /** @description Whether the project is used as template, premium feature */
            template?: boolean | null;
            /** @description Template ID */
            template_id?: number | null;
            /** @description Total number of projects found */
            total_count?: number;
            /** @description Workspace ID legacy field */
            wid?: number;
            /** @description Workspace default currency (raw, unresolved value) */
            workspace_default_currency?: string | null;
            /** @description Workspace ID */
            workspace_id?: number;
        };
        "github_com_toggl_toggl_api_internal_models.TimeEntry": {
            /** @description When was last updated, the field follows the ISO 8601 format (HH:mm:ss.ssssssZ) */
            at?: string;
            /** @description Whether the time entry is marked as billable */
            billable?: boolean;
            /** @description Related entities meta fields - if requested */
            client_id?: number;
            client_name?: string;
            /** @description Time Entry description, null if not provided at creation/update */
            description?: string | null;
            /** @description Time entry duration. For running entries should be negative, preferable -1 */
            duration?: number;
            /** @description Used to create a TE with a duration but without a stop time, this field is deprecated for GET endpoints where the value will always be true. */
            duronly?: boolean;
            /** @description Work expenses */
            expense_ids?: number[];
            /** @description Time Entry ID */
            id?: number;
            /** @description The external ID of the linked entity in the external system. */
            integration_ext_id?: string;
            /** @description The external type of the linked entity in the external system. */
            integration_ext_type?: string;
            /** @description The provider that has an entity linked to this Toggl Track time entry. */
            integration_provider?: components["schemas"]["models.IntegrationProvider"];
            /** @description Permission list */
            permissions?: string[];
            /** @description Project ID, legacy field */
            pid?: number;
            project_active?: boolean;
            project_billable?: boolean;
            project_color?: string;
            /** @description Project ID. Can be null if project was not provided or project was later deleted */
            project_id?: number | null;
            project_name?: string;
            /** @description Indicates who the time entry has been shared with */
            shared_with?: components["schemas"]["models.TimeEntrySharedWith"][];
            /** @description Start time in UTC */
            start?: string;
            /** @description Stop time in UTC, can be null if it's still running or created with "duration" and "duronly" fields */
            stop?: string;
            /** @description Tag IDs, null if tags were not provided or were later deleted */
            tag_ids?: number[];
            /** @description Tag names, null if tags were not provided or were later deleted */
            tags?: string[];
            task_active?: boolean;
            /** @description Task ID. Can be null if task was not provided or project was later deleted */
            task_id?: number | null;
            task_name?: string;
            /** @description Task ID, legacy field */
            tid?: number;
            /** @description Time Entry creator ID, legacy field */
            uid?: number;
            user_avatar_url?: string;
            /** @description Time Entry creator ID */
            user_id?: number;
            user_name?: string;
            /** @description Workspace ID, legacy field */
            wid?: number;
            /** @description Workspace ID */
            workspace_id?: number;
        };
// ...
    };
// ...
}